ci(appveyor): remove the windows ci pipeline running on appveyor
try to figure out the env variables and if conda exists

try to locate the conda executable inside a miniconda installation

ci(appveyor): remove the windows ci pipeline running on appveyor

Previously, the windows jobs that executed the so_magic test suite, where failing on the appveyor
server, since we could not install the somoclu dependency. Now we deactivate the appveyor CI and
open an issue to remember to reactivate it. We know that somoclu can run on windows, but needs to be
compiled and installed "manually", since there is no conda "recipe" (as in Linux).

activate circleci jobs