feat(tabular_operations): allow both method & classmethod as implementation of TabularData operation
Previously, the client code had to implement the PDTabularRetriever, PDTabularIterator and
PDTabularMutator interfaces as classes with (instance) methods to satisfy the interface contract.
Now, the client code can implement a required method as a class method (using the @classmethod
decorator). In other words now it is allowed to have both instance and class methods per interface
method defined.