refactor: replace doulbe inheritance with composition, define singleton pattern as type
Previously, codecay reported a complexity of 18 for the test_phi unit-test file. This commit
attempts to reduce this number. During development, certain flaws in the code were discovered and
were refactored: Phis class now is defined with a composition instead of a double inheritance, phis
are registered in a single place at runtime instead of two.